requirements.txt
# PaperForge AI dependencies

# Web framework
flask>=3.0.0
werkzeug>=3.0.0

# AI Providers
openai>=1.0.0
google-genai>=1.0.0

# PDF parsing
PyMuPDF>=1.23.0

# Utilities
typing-extensions>=4.0.0
python-dotenv>=1.0.0

# Note: Install with: pip install -r requirements.txt
